requirements.txt
streamlit==0.70.0
numpy==1.19.3
matplotlib==3.3.2
pillow==8.0.1
opencv-python==4.4.0.46
tensorflow==2.0.0
gunicorn==20.0.4
PyOpenGL==3.1.5
